order , payment service communicated , some POJO added at payment common even used lombok. both the tables are getting values from api

+++ spb_ms_javatechie/payment-service/src/main/java/com/webfullstackdev/javatechie/spbms/Entity/Payment.java
@@ -25,7 +25,7 @@ public class Payment {
 	private String paymentStatus;
 	private String transactionId;
 	private int orderId;
-	private int amount;
+	private double amount;
 
 
 	public void setTransactionId(String transactionId) {
